Move DataSources to files
